Simplify names
