+++ superstats/transition/deterministic/__init__.py
@@ -1,6 +1,6 @@
 """Deterministic transition models for latent time-varying parameters."""
 
-from .deterministic_transition import DeterministicTransition
+from deterministic import DeterministicTransition
 from .linear import Linear
 from .polynomial import Polynomial
