+++ superstats/transition/deterministic/deterministic_transition.py
