+++ superstats/transition/stochastic/jump.py
@@ -4,7 +4,7 @@ from typing import Tuple, Dict, Any
 import numpy as np
 from numba import njit, prange
 
-from .stochastic_transition import StochasticTransition, Prior
+from stochastic import StochasticTransition, Prior
 from superstats.utils.transformations import scaled_sigmoid
