+++ superstats/transition/stochastic/kernel/__init__.py
