+++ superstats/transition/stochastic/kernel/kernel.py
