+++ superstats/transition/stochastic/kernel/linear.py
