+++ superstats/transition/stochastic/kernel/periodic.py
