+++ superstats/transition/stochastic/kernel/rbf.py
