+++ superstats/transition/stochastic/kernel/registry.py
