+++ superstats/transition/stochastic/levy_flight.py
@@ -5,7 +5,7 @@ from collections.abc import Sequence
 import numpy as np
 from numba import njit, prange
 
-from .stochastic_transition import StochasticTransition, Prior
+from stochastic import StochasticTransition, Prior
 from superstats.utils.transformations import scaled_sigmoid
