+++ superstats/transition/stochastic/ornstein_uhlenbeck.py
@@ -1,39 +1,41 @@
-"""Random-walk transition model."""
+"""Ornstein-Uhlenbeck transition models."""
 
-from typing import Dict, Any
-from collections.abc import Sequence
+from typing import Tuple, Dict, Any
 import numpy as np
 from numba import njit, prange
 
-from .stochastic_transition import StochasticTransition, Prior
+from stochastic import StochasticTransition, Prior
 from superstats.utils.transformations import scaled_sigmoid
 
 
 @njit(parallel=True, fastmath=True)
-def _sample_random_walk(
+def _sample_ou(
     local_params: np.ndarray,
+    mu: np.ndarray,
+    theta: np.ndarray,
     sigma: np.ndarray,
-    delta: np.ndarray,
     bounds: np.ndarray,
 ) -> np.ndarray:
-    """Vectorized random-walk rollout across a batch, filled in place.
+    """Vectorized Ornstein-Uhlenbeck rollout across a batch, filled in place.
 
     Parameters
     ----------
     local_params : np.ndarray of shape (batch_size, steps)
         Pre-allocated trajectory array; `local_params[:, 0]` must already
         hold the initial state. Overwritten in place with the full rollout.
+    mu           : np.ndarray of shape (batch_size,)
+        Long-run mean to revert towards, per trajectory.
+    theta        : np.ndarray of shape (batch_size,)
+        Mean-reversion speed, per trajectory.
     sigma        : np.ndarray of shape (batch_size,)
-        Standard deviation of the Gaussian increments, per trajectory.
-    delta        : np.ndarray of shape (batch_size,)
-        Additive drift term, per trajectory.
+        Diffusion scale, per trajectory.
     bounds       : np.ndarray of shape (2,)
         (lower, upper) bounds passed to `scaled_sigmoid`.
 
     Returns
     -------
     local_params : np.ndarray of shape (batch_size, steps) - the same
-        array, filled with the bounded random-walk rollout
+        array, filled with the bounded OU rollout
     """
     batch_size, steps = local_params.shape
     lower, upper = bounds[0], bounds[1]
@@ -41,40 +43,47 @@ def _sample_random_walk(
     noise = np.random.randn(batch_size, steps - 1)
 
     for b in prange(batch_size):
-        increments = delta[b] + sigma[b] * noise[b]
-        local_params[b, 1:] = local_params[b, 0] + np.cumsum(increments)
+        x_prev = local_params[b, 0]
+
+        for t in range(1, steps):
+            x_prev = x_prev + theta[b] * (mu[b] - x_prev) + sigma[b] * noise[b, t - 1]
+            local_params[b, t] = x_prev
+
         local_params[b, :] = scaled_sigmoid(local_params[b, :], lower, upper)
 
     return local_params
 
 
 @njit
-def _one_step_random_walk(
+def _one_step_ou(
     x: float,
+    mu: float,
+    theta: float,
     sigma: float,
-    delta: float,
 ) -> float:
-    """Advance a single random-walk state by one step (scalar, JIT-compiled).
+    """Advance a single Ornstein-Uhlenbeck state by one step (scalar, JIT-compiled).
 
     Parameters
     ----------
     x     : float
         Previous latent state.
+    mu    : float
+        Long-run mean to revert towards.
+    theta : float
+        Mean-reversion speed.
     sigma : float
-        Standard deviation of the Gaussian increment.
-    delta : float
-        Additive drift term.
+        Diffusion scale.
 
     Returns
     -------
     x_next : float - the next latent state
     """
     noise = np.random.randn()
-    return x + delta + sigma * noise
+    return x + theta * (mu - x) + sigma * noise
 
 
-class RandomWalk(StochasticTransition):
-    """Random walk transition with Gaussian noise and optional drift.
+class OrnsteinUhlenbeck(StochasticTransition):
+    """Ornstein-Uhlenbeck mean-reverting transition.
 
     Parameters
     ----------
@@ -83,35 +92,38 @@ class RandomWalk(StochasticTransition):
     initial_prior : Prior or None, optional, default: None
         Prior for the initial latent state.
     sigma         : float or Prior or None, optional, default: None
-        Standard deviation of the Gaussian increments.
-    delta         : float or Prior, optional, default: 0.0
-        Additive drift term.
+        Diffusion scale.
+    mu            : float or Prior or None, optional, default: None
+        Long-run mean to revert towards.
+    theta         : float or Prior or None, optional, default: None
+        Mean-reversion speed.
 
     Notes
     -----
-    The `sample` method returns a dict with keys `local_params`,
-    `hyper_params` and `fixed_params`. Use `sample_one_step` to advance
-    a single time-step given numeric params.
+    Implements an OU process:
+    x_t = x_{t-1} + theta * (mu - x_{t-1}) + sigma * eps_t.
     """
 
     def __init__(
         self,
-        bounds: Sequence[float, float] | None = None,
+        bounds: Tuple[float, float] | None = None,
         initial_prior: Prior | None = None,
         sigma: float | Prior | None = None,
-        delta: float | Prior | None = None,
+        mu: float | Prior | None = None,
+        theta: float | Prior | None = None,
     ):
         super().__init__(bounds, initial_prior)
 
         self.hyper_specs = {
             "sigma": sigma,
-            "delta": delta,
+            "mu": mu,
+            "theta": theta,
         }
 
-        self.transition_name = "rw"
+        self.transition_name = "ou"
 
     def sample(self, batch_size: int, num_steps: int) -> Dict[str, Any]:
-        """Draw `batch_size` random-walk trajectories of length `num_steps`.
+        """Draw `batch_size` Ornstein-Uhlenbeck trajectories of length `num_steps`.
 
         Parameters
         ----------
@@ -135,15 +147,21 @@ class RandomWalk(StochasticTransition):
         else:
             sigma = np.full(batch_size, fixed["sigma"], dtype=self.dtype)
 
-        if "delta" in hyper:
-            delta = hyper["delta"]
+        if "mu" in hyper:
+            mu = hyper["mu"]
         else:
-            delta = np.full(batch_size, fixed["delta"], dtype=self.dtype)
+            mu = np.full(batch_size, fixed["mu"], dtype=self.dtype)
 
-        local_params = _sample_random_walk(
+        if "theta" in hyper:
+            theta = hyper["theta"]
+        else:
+            theta = np.full(batch_size, fixed["theta"], dtype=self.dtype)
+
+        local_params = _sample_ou(
             local_params,
+            mu.astype(self.dtype),
+            theta.astype(self.dtype),
             sigma.astype(self.dtype),
-            delta.astype(self.dtype),
             self.bounds,
         )
 
@@ -154,23 +172,25 @@ class RandomWalk(StochasticTransition):
         }
 
     def sample_one_step(self, x: float, params: Dict[str, Any]) -> float:
-        """Advance a single step of the random walk.
+        """Advance a single OU step.
 
         Parameters
         ----------
         x      : float
             Previous latent state.
         params : dict
-            Expected keys: `sigma`, `delta`.
+            Expected keys: `mu`, `theta`, `sigma`.
 
         Returns
         -------
         x_next : float - the next latent state
         """
+        mu = float(params["mu"])
+        theta = float(params["theta"])
         sigma = float(params["sigma"])
-        delta = float(params["delta"])
-        return _one_step_random_walk(
+        return _one_step_ou(
             x,
+            mu,
+            theta,
             sigma,
-            delta,
         )
