+++ superstats/transition/stochastic/random_walk.py
@@ -1,22 +1,22 @@
-"""Autoregressive transition models."""
+"""Random-walk transition model."""
 
-from typing import Tuple, Dict, Any
+from typing import Dict, Any
+from collections.abc import Sequence
 import numpy as np
 from numba import njit, prange
 
-from .stochastic_transition import StochasticTransition, Prior
+from stochastic import StochasticTransition, Prior
 from superstats.utils.transformations import scaled_sigmoid
 
 
 @njit(parallel=True, fastmath=True)
-def _sample_ar1(
+def _sample_random_walk(
     local_params: np.ndarray,
     sigma: np.ndarray,
-    phi: np.ndarray,
     delta: np.ndarray,
     bounds: np.ndarray,
 ) -> np.ndarray:
-    """Vectorized AR(1) rollout across a batch, filled in place.
+    """Vectorized random-walk rollout across a batch, filled in place.
 
     Parameters
     ----------
@@ -24,18 +24,16 @@ def _sample_ar1(
         Pre-allocated trajectory array; `local_params[:, 0]` must already
         hold the initial state. Overwritten in place with the full rollout.
     sigma        : np.ndarray of shape (batch_size,)
-        Noise standard deviation.
-    phi          : np.ndarray of shape (batch_size,)
-        Autoregressive coefficient.
+        Standard deviation of the Gaussian increments, per trajectory.
     delta        : np.ndarray of shape (batch_size,)
-        Additive drift term.
+        Additive drift term, per trajectory.
     bounds       : np.ndarray of shape (2,)
         (lower, upper) bounds passed to `scaled_sigmoid`.
 
     Returns
     -------
     local_params : np.ndarray of shape (batch_size, steps) - the same
-        array, filled with the bounded AR(1) rollout
+        array, filled with the bounded random-walk rollout
     """
     batch_size, steps = local_params.shape
     lower, upper = bounds[0], bounds[1]
@@ -43,34 +41,27 @@ def _sample_ar1(
     noise = np.random.randn(batch_size, steps - 1)
 
     for b in prange(batch_size):
-        x_prev = local_params[b, 0]
-
-        for t in range(1, steps):
-            x_prev = phi[b] * x_prev + delta[b] + sigma[b] * noise[b, t - 1]
-            local_params[b, t] = x_prev
-
+        increments = delta[b] + sigma[b] * noise[b]
+        local_params[b, 1:] = local_params[b, 0] + np.cumsum(increments)
         local_params[b, :] = scaled_sigmoid(local_params[b, :], lower, upper)
 
     return local_params
 
 
 @njit
-def _one_step_ar1(
+def _one_step_random_walk(
     x: float,
     sigma: float,
-    phi: float,
     delta: float,
 ) -> float:
-    """Advance a single AR(1) state by one step.
+    """Advance a single random-walk state by one step (scalar, JIT-compiled).
 
     Parameters
     ----------
     x     : float
         Previous latent state.
     sigma : float
-        Noise standard deviation.
-    phi   : float
-        Autoregressive coefficient.
+        Standard deviation of the Gaussian increment.
     delta : float
         Additive drift term.
 
@@ -79,11 +70,11 @@ def _one_step_ar1(
     x_next : float - the next latent state
     """
     noise = np.random.randn()
-    return phi * x + delta + sigma * noise
+    return x + delta + sigma * noise
 
 
-class AutoRegression(StochasticTransition):
-    """AR(1) autoregressive transition.
+class RandomWalk(StochasticTransition):
+    """Random walk transition with Gaussian noise and optional drift.
 
     Parameters
     ----------
@@ -92,37 +83,35 @@ class AutoRegression(StochasticTransition):
     initial_prior : Prior or None, optional, default: None
         Prior for the initial latent state.
     sigma         : float or Prior or None, optional, default: None
-        Standard deviation of the noise.
-    phi           : float or Prior or None, optional, default: None
-        Autoregressive coefficient.
+        Standard deviation of the Gaussian increments.
     delta         : float or Prior, optional, default: 0.0
         Additive drift term.
 
     Notes
     -----
-    Implements an AR(1): x_t = phi * x_{t-1} + delta + sigma * eps_t.
+    The `sample` method returns a dict with keys `local_params`,
+    `hyper_params` and `fixed_params`. Use `sample_one_step` to advance
+    a single time-step given numeric params.
     """
 
     def __init__(
         self,
-        bounds: Tuple[float, float] | None = None,
+        bounds: Sequence[float, float] | None = None,
         initial_prior: Prior | None = None,
         sigma: float | Prior | None = None,
-        phi: float | Prior | None = None,
         delta: float | Prior | None = None,
     ):
         super().__init__(bounds, initial_prior)
 
         self.hyper_specs = {
             "sigma": sigma,
-            "phi": phi,
             "delta": delta,
         }
 
-        self.transition_name = "ar1"
+        self.transition_name = "rw"
 
     def sample(self, batch_size: int, num_steps: int) -> Dict[str, Any]:
-        """Draw `batch_size` AR(1) trajectories of length `num_steps`.
+        """Draw `batch_size` random-walk trajectories of length `num_steps`.
 
         Parameters
         ----------
@@ -146,20 +135,14 @@ class AutoRegression(StochasticTransition):
         else:
             sigma = np.full(batch_size, fixed["sigma"], dtype=self.dtype)
 
-        if "phi" in hyper:
-            phi = hyper["phi"]
-        else:
-            phi = np.full(batch_size, fixed["phi"], dtype=self.dtype)
-
         if "delta" in hyper:
             delta = hyper["delta"]
         else:
             delta = np.full(batch_size, fixed["delta"], dtype=self.dtype)
 
-        local_params = _sample_ar1(
+        local_params = _sample_random_walk(
             local_params,
             sigma.astype(self.dtype),
-            phi.astype(self.dtype),
             delta.astype(self.dtype),
             self.bounds,
         )
@@ -171,26 +154,23 @@ class AutoRegression(StochasticTransition):
         }
 
     def sample_one_step(self, x: float, params: Dict[str, Any]) -> float:
-        """Advance a single AR(1) step.
+        """Advance a single step of the random walk.
 
         Parameters
         ----------
         x      : float
             Previous latent state.
         params : dict
-            Expected keys: `sigma`, `phi`, `delta`.
+            Expected keys: `sigma`, `delta`.
 
         Returns
         -------
         x_next : float - the next latent state
         """
         sigma = float(params["sigma"])
-        phi = float(params["phi"])
         delta = float(params["delta"])
-
-        return _one_step_ar1(
+        return _one_step_random_walk(
             x,
             sigma,
-            phi,
             delta,
         )
