+++ superstats/transition/stochastic/stochastic_transition.py
