+++ tests/test_joint_prior.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from superstats.prior import JointPrior, Prior
-from superstats.transition.stochastic_transitions import RandomWalk
+from superstats.transition.stochastic import RandomWalk
 
 BATCH_SIZE = 6
 NUM_STEPS = 10
